Different samplerates: diskrate, playrate and metarate
To make it possible to decode a song to a samplerate of E.g 22050 Hz,
whilst playing it through a high end soundcard at a samplerate of
96000 Hz, while storing the main information at a neutral 44100 Hz, we
seperated the old WAVRATE variable into diskrate, playrate and
metarate. I also added postfixes to quite a number of variables to
notify at what rate they are expressed. E.g: cue_metarate (since it is
stored on disk), x_diskrate (expressing the position in the file on
disk) and y_playrate, expressing the timeposition in the current song
at the current playrate. These variable postfixes are carried
throughout the program as to make normalisation better understandable.

The DSP drivers obtained an extra field, called playrate. When
initializing the driver should try to reach that rate.
When it cant it should set the playrate variable to the actual
playrate, so that we don't play the song too fast.

The analyzers should all still work as long as the decoding to disk
happens at 44100 Hz. Other decoding rates have not yet been tested,
but at first glance certainly seem to affect the beatgraph layout.

Bug #1185: dsp alsa latency could return nonsense
Bug #1279: rescaling of scrambler images